Backend comunicando com API

+++ app/src/main/java/com/br/everis/financialgoal/utils/validators/FieldValidator.kt
@@ -2,8 +2,7 @@ package com.br.everis.financialgoal.utils.validators
 
 import android.text.TextUtils
 import android.util.Patterns
-import java.math.BigDecimal
-import java.text.DecimalFormat
+import android.view.View
 
 class FieldValidator {
 
@@ -21,8 +20,8 @@ class FieldValidator {
 
     fun isValidPeriod(month: Int) : Boolean = month > 0
 
-    fun isValidTax(tax: Float) : Boolean = tax > 0.0
+    fun isValidTax(tax: Float) : Boolean = tax > 0F
 
-    fun isValidUniqueApplication(app: Float) : Boolean = app > 0.0
+    fun isValidUniqueApplication(app: Double) : Boolean = app > 0.0
 
 }
